+++ src/main/java/com/gogym/member/entity/BanNickname.java
@@ -2,20 +2,16 @@ package com.gogym.member.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+import com.gogym.common.entity.BaseIdEntity;
 
 @Entity
 @Table(name = "ban_nickname")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-@Builder
-public class BanNickname {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "ban_nickname_id")
-  private Long id;
+public class BanNickname extends BaseIdEntity {
 
   @Column(name = "banned_nickname", nullable = false)
   private String bannedNickname;
+
 }
